优化游戏条目结构体，添加游戏独立设置字段；实现获取最近玩的游戏列表功能；更新游戏页面逻辑以支持新功能

+++ src/core/Tools.cpp
@@ -160,7 +160,7 @@ uint32_t crc32(const std::string& path)
 
     while (file.read(buffer, sizeof(buffer)) || file.gcount())
     {
-        for (size_t i = 0; i < file.gcount(); i++)
+        for (long long int i = 0; i < file.gcount(); i++)
         {
             crc ^= (uint8_t)buffer[i];
             for (int j = 0; j < 8; j++)
